+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 ### Added
  * Basic pool and vdev health check
    * Changes to `sudoers` config on monitored system(s) may be requried
+ * (Should) model a pool's spare devices
+
+### Changed
+ * Cache, log, and spare devices are now their own component types rather than Root vDevs
 
 ## [0.7.4] - 2016-10-08
 ### Changed